+++ tests/Jroc.Test262.Tests/built-ins/Map/prototype/values/ExecutionTests.cs
@@ -38,4 +38,7 @@ public class ExecutionTests : DiskExecutionTestsBase
     public Task values()
         => ExecutionTestFromFile("values");
 
+    [Fact(DisplayName = "name")]
+    public Task name()
+        => ExecutionTestFromFile("name");
 }