+++ tests/Jroc.Test262.Tests/built-ins/WeakSet/prototype/has/ExecutionTests.cs
@@ -25,4 +25,8 @@ public class ExecutionTests : DiskExecutionTestsBase
     [Fact(DisplayName = "this-not-object-throw-undefined")]
     public Task this_not_object_throw_undefined()
         => ExecutionTestFromFile("this-not-object-throw-undefined");
+
+    [Fact(DisplayName = "name")]
+    public Task name()
+        => ExecutionTestFromFile("name");
 }